refactor(tests): colocate component boundary test modules
