+++ src/lib/hooks/sidebar/selectionState/selectionState.test.ts
