+++ src/lib/hooks/sidebar/selectionState/selectionState.ts
